+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 2.0.0
+- Added new method `waitForChange` which allows you to wait for a single change on a given directory with an optional timeout and predicate
+- `stopAll` now takes a `directory` argument to only stop watchers on a specific directory
+- Changed how `stop` and `stopAll` works
+- Exposed the weak cache of watchers
+
 # 1.1.0
 - Added a new `rename` event which fires when a file is renamed
 - Added a new method `fs_watcher.stopAll` which stops all running watchers